Add link to About page from landing page ; adjust styling to scope just to Carousel component; rename About components to AboutPage and move into pages directory
