+++ src/pages/App/App.jsx
@@ -8,7 +8,7 @@ import TextPage from '../TextPage/TextPage'
 import DemoTextPage from '../../demo/demo-pages/DemoTextPage/DemoTextPage'
 import NavBar from '../../components/NavBar/NavBar'
 import DashboardPage from '../DashboardPage/DashboardPage'
-import About from '../../components/About/About'
+import About from '../AboutPage/AboutPage'
 
 function App() {
   const [user, setUser] = useState(getUser())
@@ -40,34 +40,34 @@ function App() {
   return (
     
     <main className='App'>
-      <About />
+      { user ?
+      <>
+        <NavBar user={user} setUser={setUser} />
+        <Routes>
+          <Route 
+            path='/dashboard' 
+            element={<DashboardPage user={user} texts={texts} setTexts={setTexts} />}  />
+          <Route 
+            path='/read/:textId' 
+            element={<TextPage getText={getText} updateText={updateText} />} />
+        </Routes>
+      </>
+        :
+        <>
+          <Routes>
+            <Route 
+              path='/demo'
+              element={<DemoTextPage />} />
+            <Route 
+              path='/about'
+              element={<About />} />
+            <Route 
+              path='/*' 
+              element={<AuthPage setUser={setUser} />}  />
+          </Routes>
+        </>
+      }
     </main>
-    // <main className='App'>
-    //   { user ?
-    //   <>
-    //     <NavBar user={user} setUser={setUser} />
-    //     <Routes>
-    //       <Route 
-    //         path='/dashboard' 
-    //         element={<DashboardPage user={user} texts={texts} setTexts={setTexts} />}  />
-    //       <Route 
-    //         path='/read/:textId' 
-    //         element={<TextPage getText={getText} updateText={updateText} />} />
-    //     </Routes>
-    //   </>
-    //     :
-    //     <>
-    //       <Routes>
-    //         <Route 
-    //           path='/demo'
-    //           element={<DemoTextPage />} />
-    //         <Route 
-    //           path='/*' 
-    //           element={<AuthPage setUser={setUser} />}  />
-    //       </Routes>
-    //     </>
-    //   }
-    // </main>
   )
 }
